Widgets: moves custom Menu to widgets
