+++ src/components/Tasks/Toolbar.tsx
@@ -6,10 +6,8 @@ import Switch from '@material-ui/core/Switch';
 import FormControlLabel from '@material-ui/core/FormControlLabel';
 import IconButton from '@material-ui/core/IconButton';
 import MoreVertIcon from '@material-ui/icons/MoreVert';
-import Menu from '@material-ui/core/Menu';
 import MenuItem from '@material-ui/core/MenuItem';
 import SortIcon from '@material-ui/icons/Sort';
-import { PopoverOrigin } from '@material-ui/core/Popover';
 
 import QuickAdd from './QuickAdd';
 
@@ -20,15 +18,7 @@ import { useSelector, useDispatch } from 'react-redux';
 import { setSettings } from '../../store/actions';
 import { StoreState } from '../../store';
 
-const anchorOrigin: PopoverOrigin = {
-  vertical: 'bottom',
-  horizontal: 'right',
-};
-
-const transferOrigin: PopoverOrigin = {
-  vertical: 'top',
-  horizontal: 'right',
-};
+import Menu from '../../widgets/Menu';
 
 interface PropsType {
   defaultCollection: EteSync.CollectionInfo;
@@ -76,9 +66,6 @@ export default function Toolbar(props: PropsType) {
           keepMounted
           open={!!sortAnchorEl}
           onClose={() => setSortAnchorEl(null)}
-          getContentAnchorEl={null}
-          anchorOrigin={anchorOrigin}
-          transformOrigin={transferOrigin}
         >
           <SortMenuItem name="smart" label="Smart" />
           <SortMenuItem name="dueDate" label="Due Date" />
@@ -99,14 +86,10 @@ export default function Toolbar(props: PropsType) {
           <MoreVertIcon />
         </IconButton>
         <Menu
-          id="simple-menu"
           anchorEl={optionsAnchorEl}
           keepMounted
           open={!!optionsAnchorEl}
           onClose={() => setOptionsAnchorEl(null)}
-          getContentAnchorEl={null}
-          anchorOrigin={anchorOrigin}
-          transformOrigin={transferOrigin}
         >
           <MenuItem>
             <FormControlLabel
